+++ client/v2/client.go
@@ -442,6 +442,9 @@ type Query struct {
 	Parameters      map[string]interface{}
 }
 
+// Params is a type alias to the query parameters.
+type Params map[string]interface{}
+
 // NewQuery returns a query object.
 // The database and precision arguments can be empty strings if they are not needed for the query.
 func NewQuery(command, database, precision string) Query {
@@ -513,7 +516,7 @@ type Result struct {
 
 // Query sends a command to the server and returns the Response.
 func (c *client) Query(q Query) (*Response, error) {
-	return c.QueryCtx(nil, q)
+	return c.QueryCtx(context.Background(), q)
 }
 
 // QueryCtx sends a command to the server and returns the Response.
@@ -588,7 +591,7 @@ func (c *client) QueryCtx(ctx context.Context, q Query) (*Response, error) {
 
 // QueryAsChunk sends a command to the server and returns the Response.
 func (c *client) QueryAsChunk(q Query) (*ChunkedResponse, error) {
-	req, err := c.createDefaultRequest(nil, q)
+	req, err := c.createDefaultRequest(context.Background(), q)
 	if err != nil {
 		return nil, err
 	}