+++ query/executor.go
@@ -111,6 +111,9 @@ func (a openAuthorizer) AuthorizeQuery(_ string, _ *influxql.Query) error { retu
 // function should be preferred over directly checking if an Authorizer is nil
 // or not.
 func AuthorizerIsOpen(a Authorizer) bool {
+	if u, ok := a.(interface{ AuthorizeUnrestricted() bool }); ok {
+		return u.AuthorizeUnrestricted()
+	}
 	return a == nil || a == OpenAuthorizer
 }
 
@@ -396,7 +399,7 @@ func (e *Executor) recover(query *influxql.Query, results chan *Result) {
 		}
 
 		if willCrash {
-			e.Logger.Error(fmt.Sprintf("\n\n=====\nAll goroutines now follow:"))
+			e.Logger.Error("\n\n=====\nAll goroutines now follow:")
 			buf := debug.Stack()
 			e.Logger.Error(fmt.Sprintf("%s", buf))
 			os.Exit(1)