+++ storage/reads/group_resultset.go
@@ -23,7 +23,7 @@ type groupResultSet struct {
 	keys    [][]byte
 	nilSort []byte
 	rgc     groupByCursor
-	km      keyMerger
+	km      KeyMerger
 
 	newCursorFn func() (SeriesCursor, error)
 	nextGroupFn func(c *groupResultSet) GroupCursor
@@ -38,7 +38,7 @@ type GroupOption func(g *groupResultSet)
 // other value
 func GroupOptionNilSortLo() GroupOption {
 	return func(g *groupResultSet) {
-		g.nilSort = nilSortLo
+		g.nilSort = NilSortLo
 	}
 }
 
@@ -48,7 +48,7 @@ func NewGroupResultSet(ctx context.Context, req *datatypes.ReadGroupRequest, new
 		req:         req,
 		agg:         req.Aggregate,
 		keys:        make([][]byte, len(req.GroupKeys)),
-		nilSort:     nilSortHi,
+		nilSort:     NilSortHi,
 		newCursorFn: newCursorFn,
 	}
 
@@ -89,13 +89,13 @@ func NewGroupResultSet(ctx context.Context, req *datatypes.ReadGroupRequest, new
 	return g
 }
 
-// nilSort values determine the lexicographical order of nil values in the
+// NilSort values determine the lexicographical order of nil values in the
 // partition key
 var (
 	// nil sorts lowest
-	nilSortLo = []byte{0x00}
+	NilSortLo = []byte{0x00}
 	// nil sorts highest
-	nilSortHi = []byte{0xff} // sort nil values
+	NilSortHi = []byte{0xff}
 )
 
 func (g *groupResultSet) Err() error { return nil }
@@ -161,7 +161,7 @@ func groupNoneNextGroup(g *groupResultSet) GroupCursor {
 		mb:   g.mb,
 		agg:  g.agg,
 		cur:  cur,
-		keys: g.km.get(),
+		keys: g.km.Get(),
 	}
 }
 
@@ -174,13 +174,13 @@ func groupNoneSort(g *groupResultSet) (int, error) {
 	}
 
 	allTime := g.req.Hints.HintSchemaAllTime()
-	g.km.clear()
+	g.km.Clear()
 	n := 0
 	row := cur.Next()
 	for row != nil {
 		if allTime || g.seriesHasPoints(row) {
 			n++
-			g.km.mergeTagKeys(row.Tags)
+			g.km.MergeTagKeys(row.Tags)
 		}
 		row = cur.Next()
 	}
@@ -195,16 +195,16 @@ func groupByNextGroup(g *groupResultSet) GroupCursor {
 		g.rgc.vals[i] = row.Tags.Get(g.keys[i])
 	}
 
-	g.km.clear()
+	g.km.Clear()
 	rowKey := row.SortKey
 	j := g.i
 	for j < len(g.rows) && bytes.Equal(rowKey, g.rows[j].SortKey) {
-		g.km.mergeTagKeys(g.rows[j].Tags)
+		g.km.MergeTagKeys(g.rows[j].Tags)
 		j++
 	}
 
 	g.rgc.reset(g.rows[g.i:j])
-	g.rgc.keys = g.km.get()
+	g.rgc.keys = g.km.Get()
 
 	g.i = j
 	if j == len(g.rows) {
@@ -246,7 +246,8 @@ func groupBySort(g *groupResultSet) (int, error) {
 			nr.SortKey = make([]byte, 0, l)
 			for _, v := range vals {
 				nr.SortKey = append(nr.SortKey, v...)
-				nr.SortKey = append(nr.SortKey, ',')
+				// separate sort key values with ascii null character
+				nr.SortKey = append(nr.SortKey, '\000')
 			}
 
 			rows = append(rows, &nr)