+++ test.sh
@@ -69,7 +69,7 @@ function run_test_docker {
 
     docker run \
          --rm=$DOCKER_RM \
-         -v "$DIR:/root/go/src/github.com/influxdata/influxdb" \
+         -v "$DIR:/root/influxdb" \
          -e "INFLUXDB_DATA_ENGINE=$INFLUXDB_DATA_ENGINE" \
          -e "GORACE=$GORACE" \
          -e "GO_CHECKOUT=$GO_CHECKOUT" \