+++ tsdb/engine/tsm1/compact_test.go
@@ -121,7 +121,7 @@ func TestCompactor_CompactFullLastTimestamp(t *testing.T) {
 
 	files, err := compactor.CompactFull([]string{f1, f2})
 	if err != nil {
-		t.Fatalf("unexpected error writing snapshot: %v", err)
+		t.Fatalf("unexpected error writing snapshot: %#v", err)
 	}
 
 	r := MustOpenTSMReader(files[0])
@@ -296,8 +296,8 @@ func TestCompactor_DecodeError(t *testing.T) {
 
 	compactor.Open()
 
-	if _, err = compactor.CompactFull([]string{f1, f2, f3}); err == nil ||
-		!strings.Contains(err.Error(), "decode error: unable to decompress block type float for key 'cpu,host=A#!~#value': unpackBlock: not enough data for timestamp") {
+	if _, err = compactor.CompactFull([]string{f1, f2, f3}); err == nil || !strings.Contains(err.Error(), "decode error: unable to decompress block type float for key 'cpu,host=A#!~#value': unpackBlock: not enough data for timestamp") {
+
 		t.Fatalf("expected error writing snapshot: %v", err)
 	}
 }